requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
python-dotenv>=1.0.1
jinja2>=3.1.4
twilio>=9.3.1
openai>=1.37.0
httpx>=0.27.0
python-multipart>=0.0.9

